+++ Inference_Data_Processing/Inference_Data_Processing_Body_Tracking_Modality/inference_data_processing_body_tracking_modality/conf.py
@@ -11,7 +11,7 @@ MAIN_FOLDER_DEFAULT = pathlib.Path(__file__).parent.parent.absolute()
 MAIN_FOLDER = config('MAIN_FOLDER', default=MAIN_FOLDER_DEFAULT)
 outputs_folder = os.path.join(MAIN_FOLDER, 'outputs')
 OUTPUTS_FOLDER = config('OUTPUTS_FOLDER', default=outputs_folder)
-EXPERIMENT_ID = config('EXPERIMENT_ID', default='development-model')
+EXPERIMENT_ID = config('EXPERIMENT_ID', default='dev_model')
 datasets_folder = os.path.join(MAIN_FOLDER, 'datasets')
 DATASETS_FOLDER = config('DATASETS_FOLDER', default=datasets_folder)
 
@@ -36,39 +36,8 @@ if os.path.exists(PATH_CUSTOM_SETTINGS):
     with open(PATH_CUSTOM_SETTINGS, 'r') as f:
         CUSTOM_SETTINGS = json.load(f)
 
-DATA_PATH = os.path.join(DATASETS_FOLDER, CUSTOM_SETTINGS["dataset_config"]["dataset_name"])
 # Define components outputs folder
 if "modality" in CUSTOM_SETTINGS["dataset_config"]:
     modality = CUSTOM_SETTINGS["dataset_config"]["modality"]
 else:
     modality = "default_modality"
-
-MODALITY_FOLDER = os.path.join(
-    OUTPUTS_FOLDER,
-    CUSTOM_SETTINGS["dataset_config"]["dataset_name"],
-    modality,
-)
-
-
-BT_LABEL_TO_EMOTION = {
-    "01": "BORED",
-    "02": "ENGAGED",
-    "03": "FRUSTRATED",
-    "04": "UNLABELED"
-}
-BT_EMOTION_TO_LABEL = {
-    "BORED": "01",
-    "ENGAGED": "02",
-    "FRUSTRATED": "03",
-    "UNLABELED": "04"
-
-}
-
-ID_TO_LABEL = {
-    0: "BORED",
-    1: "ENGAGED",
-    2: "FRUSTRATED",
-    3: "UNLABELED"
-}
-
-PUBLISHER_ON = config('PUBLISHER_ON', default=CUSTOM_SETTINGS['inference_config'].get('publisher', False), cast=bool)